### 2019-02-05 #1232 @kyleniemeyer
@arfon whedon is having trouble finding the paper, though at https://github.com/a-slide/pycoQC I see it at `paper/paper.md`.

@brainstorm @mgymrek @pjotrp @yochannah do one of you have the bandwidth to edit this submission? It looks bioinformatics-y to my untrained eyes... :)
