### 2019-02-05 #1236 @whedon
[REVIEW]: pycoQC, interactive quality control for Oxford Nanopore Sequencing
**Submitting author:** @a-slide (<a href="http://orcid.org/0000-0001-8772-6776">Adrien Leger</a>)
**Repository:** <a href="https://github.com/a-slide/pycoQC" target ="_blank">https://github.com/a-slide/pycoQC</a>
**Version:** v2.2.0
