+++ README.md
@@ -142,17 +142,16 @@ In addtion for guppy barcoded datasets you also have to provide a barcoding summ
 `pycoQC -f {path to summary_sequencing file} -b {path to barcoding_summary file} -o {path to output html report file}`
 
 
+Example HTML reports generated by pycoQC from sample data are available below.
 
-Some example HTML reports generated by pycoQC are available below:
-
-* [Albacore-1.2.1_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Albacore-1.2.1_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Albacore-1.2.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Albacore-1.2.3_basecall-1D-RNA_sequencing_summary.txt.gz)
-* [Albacore-1.7.0_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Albacore-1.7.0_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Albacore-2.1.10_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Albacore-2.1.10_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Albacore-2.1.10_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Albacore-2.1.10_basecall-1D-RNA_sequencing_summary.txt.gz)
-* [Albacore-2.3.1_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Albacore-2.3.1_basecall-1D-RNA_sequencing_summary.txt.gz)
-* [Guppy-2.1.3_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Guppy-2.1.3_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Guppy-2.1.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Guppy-2.1.3_basecall-1D-RNA_sequencing_summary.txt.gz)
+* [Albacore-1.2.1_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Albacore-1.2.1_basecall-1D-DNA_pycoQC.html)
+* [Albacore-1.2.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Albacore-1.2.3_basecall-1D-RNA_pycoQC.html)
+* [Albacore-1.7.0_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Albacore-1.7.0_basecall-1D-DNA_pycoQC.html)
+* [Albacore-2.1.10_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Albacore-2.1.10_basecall-1D-DNA_pycoQC.html)
+* [Albacore-2.1.10_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Albacore-2.1.10_basecall-1D-RNA_pycoQC.html)
+* [Albacore-2.3.1_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Albacore-2.3.1_basecall-1D-RNA_pycoQC.html)
+* [Guppy-2.1.3_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Guppy-2.1.3_basecall-1D-DNA_pycoQC.html)
+* [Guppy-2.1.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Guppy-2.1.3_basecall-1D-RNA_pycoQC.html)
 
 pycoQC help can be obtained by using `pycoQC --help`.
 
