+++ docs/index.md
@@ -1,18 +1,17 @@
 # pycoQC Documentation Index
 
-### Jupyter notebook
+### Jupyter notebook API example
 
 * [Live Jupyter Notebook](https://mybinder.org/v2/gh/a-slide/pycoQC/master?filepath=docs%2FpycoQC_usage.ipynb)
 * [Static Jupyter Notebook](https://a-slide.github.io/pycoQC/pycoQC_usage.html)
 
-###  Command line interface
+###  Command line interface output example
 
-* [Albacore-1.2.1_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Albacore-1.2.1_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Albacore-1.2.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Albacore-1.2.3_basecall-1D-RNA_sequencing_summary.txt.gz)
-* [Albacore-1.7.0_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Albacore-1.7.0_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Albacore-2.1.10_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Albacore-2.1.10_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Albacore-2.1.10_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Albacore-2.1.10_basecall-1D-RNA_sequencing_summary.txt.gz)
-* [Albacore-2.3.1_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Albacore-2.3.1_basecall-1D-RNA_sequencing_summary.txt.gz)
-* [Guppy-2.1.3_basecall-1D-DNA](https://a-slide.github.io/pycoQC/summary/Guppy-2.1.3_basecall-1D-DNA_sequencing_summary.txt.gz)
-* [Guppy-2.1.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/summary/Guppy-2.1.3_basecall-1D-RNA_sequencing_summary.txt.gz)
-
+* [Albacore-1.2.1_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Albacore-1.2.1_basecall-1D-DNA_pycoQC.html)
+* [Albacore-1.2.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Albacore-1.2.3_basecall-1D-RNA_pycoQC.html)
+* [Albacore-1.7.0_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Albacore-1.7.0_basecall-1D-DNA_pycoQC.html)
+* [Albacore-2.1.10_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Albacore-2.1.10_basecall-1D-DNA_pycoQC.html)
+* [Albacore-2.1.10_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Albacore-2.1.10_basecall-1D-RNA_pycoQC.html)
+* [Albacore-2.3.1_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Albacore-2.3.1_basecall-1D-RNA_pycoQC.html)
+* [Guppy-2.1.3_basecall-1D-DNA](https://a-slide.github.io/pycoQC/data/output/Guppy-2.1.3_basecall-1D-DNA_pycoQC.html)
+* [Guppy-2.1.3_basecall-1D-RNA](https://a-slide.github.io/pycoQC/data/output/Guppy-2.1.3_basecall-1D-RNA_pycoQC.html)
