+++ paper/paper.bib
@@ -79,7 +79,7 @@
 	url = {https://github.com/GenomicParisCentre/toulligQC},
 	urldate = {2019-01-18},
 	publisher = {GenomicParisCentre},
-	author = {{Bérengère Laffay} and {Lionel Ferrato-Berberian} and {Laurent Jourdren} and {Sophie Lemoine} and {Stéphane Le Crom}},
+	author = {Laffay, Bérengère and Ferrato-Berberian, Lionel and Jourdren, Laurent and Lemoine, Sophie and Le Crom, Stéphane},
 	month = oct,
 	year = {2018},
 	note = {original-date: 2017-05-15T14:41:53Z}
@@ -89,7 +89,7 @@
 	title = {The {Jupyter} {Notebook}},
 	url = {https://jupyter-notebook.readthedocs.io/en/stable/},
 	urldate = {2019-01-18},
-	author = {{Jupyter Project and Community}},
+	author = {{Jupyter Project}},
 	month = jan,
 	year = {2019}
 }
@@ -110,7 +110,7 @@
 @misc{plotly_technologies_inc_plotly_2015,
 	title = {Plotly, charting tool for online collaborative data science},
 	url = {https://plot.ly},
-	author = {{Plotly Technologies Inc}},
+	author = {{Plotly Technologies}},
 	year = {2015}
 }
 
