+++ paper/paper.md
@@ -33,7 +33,7 @@ The electrical signal acquired by the array of nanopores is stored in hdf5 forma
 
 Here we present pycoQC, a new tool to generate interactive quality control metrics and plots from basecalled nanopore reads or summary files generated by Albacore and Guppy. Although there are other open-source alternatives such as Nanoplot [@de_coster_nanopack_2018], MinionQC [@lanfear_minionqc_2018] and toulligQC [@berengere_laffay_toulligqc_2018], pycoQC has several novel features:
 
-* Integration with the plotly [@plotly_technologies_inc_plotly_2015] python charting library to create dynamic D3.js visualizations.
+* Integration with the plotly python charting library to create dynamic D3.js visualizations [@plotly_technologies_inc_plotly_2015].
 * Extensive python API developed for interactive data exploration in Jupyter Notebooks [@jupyter_project_and_community_jupyter_2019] ([example notebook](https://a-slide.github.io/pycoQC/pycoQC_usage.html)).
 * Simple command line interface to generate customizable interactive HTML reports ([example report](https://a-slide.github.io/pycoQC/data/output/Albacore-2.3.1_basecall-1D-RNA_pycoQC.html)).
 * Multiprocessing Fast5 feature extraction program to generate a summary file directly from Fast5 files.
