### 2019-02-06 b5bf3f5
Add alternative packages list and link to comparison
