+++ README.md
@@ -277,6 +277,18 @@ optional arguments:
                         (default: 0)
 ```
 
+## Alternative and complementary packages
+
+Here is a non-exhaustive list of alternative/complementary packages that could also be used to QC Nanopore sequencing data:
+
+* [Nanoplot](https://github.com/wdecoster/NanoPlot): Generic QC. Part of NanoPack metapackage which offers a broad range of functionalities. *(Python)*
+* [MinION-QC](https://github.com/roblanf/minion_qc): Generic QC with additional interesting plots such as `flowcell_overview`. *(R)*
+* [toullig-QC](https://github.com/GenomicParisCentre/toulligQC): Generic QC. Requires additional input files but gives base resolution information *(Python)*
+* [flowcellvis](https://github.com/mattloose/flowcellvis): Animated GIF of flow cell performance. *(Python)*
+* [Basic QC Tutorial](https://github.com/nanoporetech/ont_tutorial_basicqc): Provided by ONT. Step by step hands-on guide to perform a basic quality control of sequencing data. *(R)*
+
+pycoQC results are consistent with other packages. Reports of the same datasets generated by Nanoplot, MinIONQC and pycoQC are available at the following URL: https://www.ebi.ac.uk/~aleg/data/pycoQC_test/comparison/
+
 ## Contributing
 
 Thanks for considering contributing to our_package!
