### 2019-02-06 4c5ec13
Add bleeding edge clone option
