### 2019-02-06 #1236 @a-slide
Hi @pjotrp and @george-githinji,

I confirm that I can tick all the boxes, except **automated tests**. Indeed, there is no straightforward way to do it as the outputs of pycoQC methods are dynamic plots. One possibility is to inspect visually the outputs and compare it with other software which I have done. Reports generated from the same datasets with pycoQC, Nanoplot and MinIONQC are available online at https://www.ebi.ac.uk/~aleg/data/pycoQC_test/comparison/. Although the 3 programs have different implementation, the results are similar for all 3 summary files tested. If this is an acceptable solution to you, I am happy to upload the datasets on Zenodo to get a DOI.

In addition, I made a dew minor changes in pycoQC and updated the software version to 2.2.1.

Thanks
