### 2019-02-07 #1236 @pjotrp
Thanks. I am fine with not having automated tests for visualisations. One improvement is to create regresssion tests (where you just compare the outputs). But I think that is for more complex systems.

@george-githinji you can start review.
