### 2019-02-11 #1236 @george-githinji
I am currently going through the package and the the paper submission. Should get back to you in 2 days. Thanks!
