### 2019-02-22 #1236 @george-githinji
The authors provide a tool for summarizing nanopore sequence data based on the sequence summary file generated by the base callers. They make use of third-party tools to provide summary statistics and plots. The tool is well documented and provides clear and standard methods for installation under open source license. They compare the tool with similar functionality and make argument for extra features. It would be great if the tool provided raw summary data used in each visualization/plot to the users so as to  allow alternative visualization methods or environments.  The dynamic plots would be great for real-time analysis during a nanopore run and might want to consider such a feature in the future. I would recommend publication to JOSS.
