### 2019-02-24 #1236 @pjotrp
@a-slide The review process is now complete. To finalize your submission and accept your paper in JOSS, we need two things. First, can you confirm that all references in your bibliography have a DOI in the bibliography (if one exists).

Second, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service.

To do so:

3. Post a comment here with the DOI for the release.

     * Get a new proof with the @whedon generate pdf command.
     * Download the proof, check all references have DOIs, follow the
       links and check the references.
     * Give the paper a proof-read and ask authors to fix typos.
     * Ask the author to make a Zenodo archive, and report the DOI in the
       Review thread.
     * Check the Zenodo deposit has the correct metadata (title and author
       list), and request the author edit it if it doesnt match the
       paper.
     * Run @whedon set <doi> as archive.
     * Run @wedon set <v1.x.x> as version if the version was updated.
     * Add the accepted label.
