### 2019-02-24 #1236 @arfon
> @a-slide The review process is now complete. To finalize your submission and accept your paper in JOSS, we need two things. First, can you confirm that all references in your bibliography have a DOI in the bibliography (if one exists).

`@whedon` can help with this by typing `@whedon check references`
