### 2019-02-26 #1236 @a-slide
Dear all,

Sorry for the late reply and many thanks for accepting our paper.

@george-githinji, We would like to thank you for your time and for the fair review.
Regarding your remark on raw data, although it is not documented all the pycoQC functions actually returns the plots in JSON format. If you don't plot it with plotly iplot function the raw data can easily be extracted from the object returned.

@arfon and @pjotrp we published a new version of pycoQC on the master branch with minor changes related to the documentation ([see relevant pull request](https://github.com/a-slide/pycoQC/commit/bedf3e308864140d15816e77d1767e58a02eaf12) ). The last version number is **[2.2.2](https://github.com/a-slide/pycoQC/tree/v2.2.2)** with DOI **[10.5281/zenodo.2578275](https://zenodo.org/record/2578275)**. The pre-compiled version hosted on pypi was also updated.
As for references, some don't have a DOI as I cited several unpublished tools or github repositories.

We hope this answers your questions and look forward for the final version of the paper.

Best
