### 2019-02-28 #1236 @danielskatz
Hi @pjotrp - the new practices, now that we have AEiCs and Arfon doesn't have to do this all himself, is to notify @openjournals/joss-eics when a paper is ready to go.  That's me this week, and I'll go ahead and do this now.
