### 2019-02-28 b4ac1f2
updating cases of names
