### 2019-02-28 #1236 @danielskatz
@a-slide - Can I ask for some small updates to the paper?  All related to cases...  I've suggested them via https://github.com/a-slide/pycoQC/pull/55
