### 2019-02-28 #1236 @a-slide
Hi @danielskatz,
Sure thanks. Should I also update the package version number ?
