### 2019-02-28 #1236 @danielskatz
I don't think it's needed for such a small text-only change in the paper
