### 2019-02-28 #1236 @tleonardi
I've also added a pull request for a problem with the @ in my affiliation
