### 2019-02-28 #1236 @danielskatz
We are aware there is a problem with the paper on the JOSS site - hopefully to be resolved soon.
